შეიცვალა MediatR ის ბიბლიოთეკა

+++ Core/Core.Application/Interfaces/Repositories/IEventRepository.cs
@@ -6,6 +6,6 @@ public interface IEventRepository
     Task SaveAsync<TAggregate>(TAggregate aggregate) where TAggregate : Aggregate, new();
 
     Task<TAggregate?> LoadAsync<TAggregate>(Guid aggregateId) where TAggregate : Aggregate, new();
-    Task<TAggregate?> LoadAsync<TAggregate>(Guid aggregateId, long version) where TAggregate : Aggregate, new();
+    Task<TAggregate?> LoadAsync<TAggregate>(Guid aggregateId, int version) where TAggregate : Aggregate, new();
     Task<TAggregate?> LoadAsync<TAggregate>(Guid aggregateId, DateTime timeStamp) where TAggregate : Aggregate, new();
 }
