+++ Core/Core.Domain/Basics/AuditableEntity.cs
@@ -5,7 +5,7 @@ public abstract class AuditableEntity : BaseEntity
     /// ჩანაწერის ცვლილების რიგითი ნომერი
     /// გვიცავს გაუთვალისწინებელი, განმეორებითი Update -ებისგან
     /// </summary>
-    public virtual long Version { get; set; } = -1;
+    public virtual int Version { get; set; } = -1;
 
     public virtual DateTime DateCreated { get; set; } = DateTime.UtcNow;
     public virtual Guid? CreatedBy { get; set; }
